+++ src/dom/RequestAnimationFrame.js
@@ -82,6 +82,17 @@ var RequestAnimationFrame = new Class({
          */
         this.lastTime = 0;
 
+        /**
+         * The target FPS rate in ms.
+         * Only used when setTimeout is used instead of RAF.
+         *
+         * @name Phaser.DOM.RequestAnimationFrame#target
+         * @type {number}
+         * @default 0
+         * @since 3.21.0
+         */
+        this.target = 0;
+
         var _this = this;
 
         /**
@@ -119,7 +130,7 @@ var RequestAnimationFrame = new Class({
         {
             var d = Date.now();
 
-            var delay = Math.max(16 + _this.lastTime - d, 0);
+            var delay = Math.min(Math.max(_this.target * 2 + _this.tick - d, 0), _this.target);
 
             _this.lastTime = _this.tick;
 
@@ -139,8 +150,9 @@ var RequestAnimationFrame = new Class({
      *
      * @param {FrameRequestCallback} callback - The callback to invoke each step.
      * @param {boolean} forceSetTimeOut - Should it use SetTimeout, even if RAF is available?
+     * @param {number} targetFPS - The target fps rate (in ms). Only used when setTimeout is used.
      */
-    start: function (callback, forceSetTimeOut)
+    start: function (callback, forceSetTimeOut, targetFPS)
     {
         if (this.isRunning)
         {
@@ -150,6 +162,8 @@ var RequestAnimationFrame = new Class({
         this.callback = callback;
 
         this.isSetTimeOut = forceSetTimeOut;
+
+        this.target = targetFPS;
 
         this.isRunning = true;